Revert "patches: add temporary mp3 decoding fix"
This reverts commit acef484c9a84c2a2701c933affbc3f67bb778791.

This wasn't ever needed since the new gas-pp version was taken
in use at the end of May. The decoding issues in the latest release
must have come from either using an older version of gas-pp by accident
or from not updating the contribs used by the rest of the build system
properly.