Fix subtitle downloads by not messing encodings
VLC was downloading subtitle files as an NSString and assumed certain
encodings. This is useless to do in the controller, because VLCKit will,
eventually, autodetect encodings and “do the right thing”.

Because of this, some subtitle files (e.g: UTF-16 encoded ones) wouldn't
work over the network, but would work if copied to the device.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>