Implement special gesture recognizer for Siri Remote which can detect the touch location and select short and long presses Both the SiriRemote recognizer and the pan gesture recognizer work with the same touches.
Possible improvements to think about:
An improvement in user experience might be to cancel the SiriRemote recognizer instead of updating the touch location if the touch moved to much.
But then we have the problem that the hint does not update to the other location until the user lifts the finger and puts it down again.