tvOS: Fixed transport bar not being updated when playback is paused
On tvOS, when the playback is paused, the transport bar is not being updated. This is because the playbackPositionUpdated callback in the VLCFullscreenMovieTVViewController class only updates the transport bar in case it is visible. This makes sense as it does not need to be updated in case it is not visible. However, in case of the playback being paused, the mediaPlayerStateChanged callback of the VLCFullscreenMovieTVViewController class, which handles showing the transport bar, is called after the last playbackPositionUpdated callback. Because of that, the transport bar is displayed but it displays the old position of when the transport bar was last hidden.

This commit adds an additional update of the transport bar in case the playback is being paused.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>
(cherry picked from commit 1dd7fde409484285d2bee8cea4630b4e007c3b8e)