Filter files displayed by the FTP server browser
Unlike other server browsers, the FTP server browser displays all files that are found in a certain folder, which includes files that cannot be played by VLC like metadata files or subtitles. This commit implements filtering of displayed files so that only playable video and audio files are displayed.

In addition, this commit also fixes a bug where when a server browser implements the subtitleURL property, the VLCServerBrowsingController would not use the URL.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>