Do not filter files on libvlc browsed servers in libvlc, but in the server browser
Currently, when listing the files in a folder that is browsed using libvlc, libvlc will filter the files in that folder to exclude files with known non-playable file extensions. This is a problem because the server browser that handles libvlc browsed servers needs to know about all files in a folder in order to be able to find subtitles in external files. This commit changes the behaviour to not filter any files in libvlc, but in the server browser instead so we are able to add support for finding external subtitle files later.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>
(cherry picked from commit f1bf14132e75b7b3e9066fae15b6b4a428c5c256)