URL decode name in VLCNetworkServerBrowserItemVLCMedia
In order to determine its name, a VLCNetworkServerBrowserItemVLCMedia will first check if its media has been parsed and use the parsed title information if it has been. If it hasn't been parsed, it will fallback to the filename of the media and if that doesn't work, the whole path of the media. The filename and path of media are URL encoded (at least in the case where the media is located on an SMB share). This commit adds decoding of the URL encoded values when setting the name of a VLCNetworkServerBrowserItemVLCMedia.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>