Migrate GTM-OAuth2 to GTMAppAuth for GDrive authentication
Google has deprecated GTM-OAuth2 and has replaced it with GTMAppAuth.
This patch migrates the use of GTM-OAuth2 library.
Additionally GTMAppAuth needs a new clientID for iOS clients.
Therefore for this patch to work correctly, kVLCGoogleDriveClientID and kVLCGoogleRedirectURI
needs to be setted in VLCGoogleDriveConstants.h as well as GoogleAuth in iOS-Info.plist

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>