VLCNetworkServerBrowserUPnP: Add multi-protocol handling (closes #55)
If a media have multiple URI with an image, the image was choosed as the
default URI. Hence, this blocked the streaming of some media and could
even lead to a crash with a big media folders.
This patch fixes temporarily this issue waiting for the integration of the
libvlc's services discovery and the de-integration of upnpx.
Additionally, adds constants of protocols.