VLCNetworkCell: Fix - file/folder name ovlapping download button
When viewing local network files there was an issue where if the folder/file name was extremely long, the text would cover the download button making it difficult to see. This fixes that issue by updating the constraints in VLCNetworkList.xib and adding the [super prepareForReuse]; call in the cell class.

Signed-off-by: Carola Nitz <nitz.carola@googlemail.com>