VLCLibraryViewController: fix disappearing Searchbar on rotation
When we setediting to no on rotation we also made the searchbar disappear.
This was done because multiple selection was not possible on collectionviews after rotation.
Instead of quick fixing it by resetting the editmode we now setup and update the tableview
and collectionview no matter if they're currently displayed or not.
That way when we switch the displaymode both views are steup correctly.