newUI: Introduction of the new UI
Move from the sidemenu to use the tabbarcontroller, removed the MenuRevealButton
Add the MediaViewController and adjust the architecture in the VLCDisplayController
Created a subclass for the  view of the VLCDisplayController to let touches go through
Have only one instance of the Playbackcontroller that is going to be part of the tabbarcontroller
Changed the code that is currently calculating the miniviews position. This still needs fixing in case you turn the phone
Add sort button still need to implement functionality
Add Presentationtheme and darkmode. Move Appearance code into own class
Move the Wificell into the VLCServerListViewController