VLCHTTPFileDownloader: only download files that are supported by VLC
This essentially reverts commit 95ad1b17e966b5dd4a26b94d8fa6daa8b3348f36
We ended up with a succesful download for a not supported format which we then didn't display in the library.
The user couldn't play or delete the downloaded file.
Instead of showing not supported files in the library, we're not allowing downloads in the first place.