Use the chapter index to look up the description
A tvOS bug exists where all chapters show the name of the first chapter,
rather than the names of each chapter. This is because the player's
current title's index is used to lookup in the chapter descriptions
array. Using the current chapter's index instead results in a valid
lookup and working chapter titles on tvOS.

Signed-off-by: Felix Paul Kühne <felix@feepk.net>