Layout the subframe so that it's shown on first visit
On tvOS, there's a bug where the list of chapters isn't shown when the
info pane is first opened. Switching to another tab of the pane and then
back to the (default, first) Chapters tab displays the subview correctly.
Calling layoutIfNeeded when binding the subview to the info controller
seems to ensure that it's ready to be displayed on the first visit.