ActionSheetCell: Modify ActionSheetCellModel to configure ActionSheetCell
ActionSheetCells are now configured by passing a ActionSheetCellModel object to a convenience initializer or
through the configure(withModel model:) method.
UISwitch added as an accessoryType and new accessoryView added to display the varying accessoryTypes.
ActionSheetCellDelegate protocol defined to allow customization of whether the cell should update its theme
colors in response to a theme change.