MediaLibrarySerivce: Remove .Trash on each foreground notification
On each foreground notification we check if there is a `.Trash` folder which is invisible
for the user that can be created by deleting media from the Files app.
This could lead to disk space issues.
For now since we do not handle restoration, we delete the `.Trash` folder every time.

Closes #656