VLCPlaybackService: State recovery: Check for media duration
In the case where the media is short (e.g. 3 sec), the last position
check wouldn't be enough to restrict the opening and closing behaviour.

Indeed, when the position is saved from the `libvlc_MediaPlayerStopped`
event. It seems that the position isn't guaranteed to be correct even at
the end of a playback.

This will match VLC Android's behaviour.