SMB: Force SMBv1
Some samba servers (on Windows 7) implement both SMBv2 and SMBv1. The problem is
that the SMBv2 part is not configured like the SMB1 one. Only SMBv1 seems to
reflect the user configuration (using Windows Settings, not anything
complicated like via powershell/regedit).

If we try to connect to such server via libsmb2, the server will return a
SMB2_STATUS_ACCESS_DENIED (0xC0000022) status. Our libsmb2 module will then ask
the user for credentials via a dialog. The problem is that no credentials will
ever work since only the SMBv1 part is configured.

This forces SMBv1 but warns and ask the user for a confirmation if this
is wanted.