VLCSharedLibraryParser: Fix potential crash on unknown selector
It was assumed that parsedContents was always an array of
NSDictionnary.

This makes sure that the object is a NSDictionnary in order to avoid a
potential crash.

Indeed in some rare cases, it seemed that the first object can be a
NSString which leads to a crash with unknown selector.