iOS: Fix color theme changes when set to automatic
This fixes the color theme changes when the setting is set to automatic
and the user changes the theme from the control center.

Indeed, before, we used a global extension of UITabBarController to control
it globally but it seems that this doesn't work as it use to.

Therefore, we extend more specific view controllers or add the needed code
inside the targetted view controller.