Launch Screen: replace gradient with solid color
On iPhone X and later, the launch screen is unhappy about full screen
images and will show black instead.

This commit replaces the gradient bitmap with a software created solid
color that matches the official VLC orange and thereby fixes #1157.

Creating a gradient in software is prevented by the OS as custom(-ized)
classes are not allowed in Launch Screens.