Dropbox: Update to API V2
This updates to the latest Dropbox API due to the deprecations that will
come on the end of September 2021.

This is for now a simple and working migration by using scoped permission and
`authorizeFromControllerV2` for the authorization flow.

Ideally, we should later on consider migrating to the Swift library and use the
new recommended authorization flow.

Closes #1151