Remote playback view: Sort files by display name
The remote playback gallery page displays files in an arbitrary order,
which makes it very difficult to navigate and find files when there are
a lot loaded on the device.

This change sorts files by their display names (final path element).
Note that sorting is done by raw String value rather than, e.g., Unicode
collation rules. While this may not be ideal for all languages, it at
least means that a given set of files will appear in a deterministic
order.