VLCTime: remove int64->int32 casts
The media duration may be longer than int32 and is already defined as
int64 in media library. This should be case throughout the app.

VLCTime is now initialized with a NSNumber object that keeps the
variable dimensions without additional casting. There is no performance
impact as VLCTime stores its internal value as a NSNumber any that is
created a tad earlier now.

Fixes #1213