playback service / metadata: update timing on major position changes
The OS automatically calculates the approximate playback position and
time based on the position and the playback rate we provide. This
calculation is off on major position changes. However, when updating
the dictionary on every position intervall, this would be very costly.
This approach tries to update it for 10 times and then gives up. This
is enough for local media and most networks.

This fixes #1224.