Apple TV server list: remove batch updates for the collection view
This functionally reverts efc27cd7 as the compare block requires
indistinguishable items. However, a UPnP server can host both SAT>IP
and multimedia which cannot be recognized from outside libvlc yet
leading to only either one of them to be displayed.

On server list change, the entire view is being reloaded now, which is
less optimal but ok given the performance of current devices. A
reimplementation should be considered as soon as servers can be cleanly
recognized in libvlc's external API.