Remove AppCenter integration in favor of Apple's native tools
The native tools provided by iOS and tvOS are sufficient now to collect
bug reports and usage statistics, so there is no longer the need to
integrate with an external service whose backend is closed source and
hosted in the US with potential privacy implications. The latter is
also true for the system provided service - however, the user consent
is managed outside our app on a global level.

Fixes #1233