UPnP: add work-around to support SAT>IP
SAT>IP needs the host address of the UPnP server set as an option as
when using a generic playlist, the host address will be 'sat.ip' and
playback will fail. According to section 3.4.1 of the SAT>IP
specification, it is required to provide an icon for a SAT>IP server
via UPnP, so it is safe to query for the actual host address as the
media's URL will point to the generic playlist.