VideoPlayerViewController: take priority over system keyboard behaviour
On recent OS we need to specify on certain keys that we want the priority
in order to receive the event correctly.

UIKeyCommand.wantsPriorityOverSystemBehavior is introduced in iOS 15 SDK but we actually still use Xcode 12.4 on CI.
This old version only provides SDK for iOS 14.4 max.
Hence we use ObjC apis to call the method and still have the ability to build with older Xcode versions.

Closes #1237

Co-authored-by: Maxime Chapelet <umxprime@videolabs.io>