movie collection view cell: remove 3 stack views and merge 2 labels
This replaces 2 stack views used to align the thumbnails for the group
appearance with the respective auto-layout guides to noticeably improve
the scrolling speed on slower devices.

This also replaces another stack view used to align the thumbnail for
the regular movie  appearance with the respective auto-layout guides
for additional speed improvements.

Finally, this merges 2 label into 1 by using a format string.