video filter : general improvements
The playback service now provide a single accessor to control the video filter.

This transfers the responsibility from the view to the service allowing to :
- access to all filter parameters' value, default value and allowed values range
- reset the video filter if needed

The view also now disable the filter properly when a reset is requested.

Those changes also fix another issues where filter was always reset when playback view was minimised or closed.

Closes #1240