VLCPlaybackService: set fullscreenSeessionRequested value to true
Since this property is mostly used for network and cloud related playback
this will set by default the player to fullscreen since it was already
what was done.

Additionnaly, this will fix the behaviour for the cloud services that
didn't specifically set the value.

The "Play video in fullscreen" settings is now respected consistently
for network and cloud.