URLHandler: remove percent encoding if the entire URL is encoded
However, we may not always decode the URL especially if only components
are encoded. This adds a logic to determine the encoding state of the
provided URL parameter.

In 2015 with VLC-iOS 2.4 we did not specify the encoding state of the
URL parameter so many competing implementations exist.

Fixes #1297 and its duplicates