Settings: fix display in Chinese when region is China (mainland)
This removes the previous error prone and slow custom way of trying to
find the correct language and region codes with consecutive read
attempts of the files.

It is replaced by a lookup of the Settings.bundle and using the system
API to just ask for the localized string, which will automatically
return the correct result.

(cherry picked from commit 7354725cc447c1d7b83e00e12fcc83d375d1aebc)