video player: seek forward by 30s
This increases the forward seek step on the swipe gesture to 30s to
match the behavior of the previous player as it often requested by
users.

We did that in 2013 as we assumed forward seeking a larger step seems
to be what most users want, while going back 10s are fine for more
granular control.

(cherry picked from commit bf5575156a8a9b4ea2caf93d65912723d8893c60)
Signed-off-by: Soomin Lee <bubu@mikan.io>