video player: respect gesture options in the user sense
Users do not differentiate between the slider and the gesture offered
by the player interface. If the gesture is disabled, so should be the
visual control that is also manipulated with a gesture.

Additionally, this fixes the slider background of the opposite slider
to appear on animation if either the volume or the brightness gesture
is disabled.

(cherry picked from commit 1754412e658b8bfce3c1f2b0df0aba0067b34f43)