PlaybackSpeedView: set a fixed height
This is done so that the view no longer overlays the subtitles when the
user is trying to set the delay.

Additionally, this removes 2 stack views with 2 additional spacer views
without changing the layout in anyway.

(cherry picked from commit 7956d8b6abc8b86266fa15d91a77c7ccdea960d8)