Add preamp setting with a default of 6 dB
This matches the previous behavior that renders audio at a level
comparable to other playback apps on iOS and tvOS. It effectively
reverts f68376ba.

When resetting an equalizer profile, the newly configured default preamp
level will also be respected.

Hereby, an option is introduced to further increase the preamp level
but also to lessen it to 0 dB and below to retain audio quality without
compression at the expense that other apps on the same device might be
louder leading to an incoherent playback experience.

libVLC allows a preamp level of up to 20 dB. This patch limits the
level to 12 dB to prevent users from breaking their device speakers and
more notably cheap Bluetooth device incapable of handling the amount of
compression.

This patch fixes #374 and #1423

(cherry picked from commit d7d01f3aa52d96a7ad72ff8831536122508b7dd4)