VLCPlaybackService: Refactor the repeat current item mode
The repeat mode allowing to repeat the current item is now switched to
the repeat all items mode when the next or previous button is pressed
and the next or previous media in the play queue starts playing.

The current item mode behaviour keeps looping the current media if the
user does not interact with the next/previous buttons.

Closes #1432