VLCPlaybackService: Fix wrong stream recognition
If the current playing media has been recognized as a live stream
and it turns out to have a positive value length, the player in the
control center is updated to the default player.

The playback is not interrupted anymore at the end of the media and
the play/pause and previous/next interactions do not lead to the end
of the playback either.

Closes #1416