Re-write App Coordinator
This allows access to all global services from anywhere within the app.

It also simplifies the handling of the media library service within its
visual representation as in virtually all cases, access to other
services is not needed at all.

Additionally (which was the reason for doing this at all), the media
library service is now available independently of any presence of the UI
which is needed for CarPlay support.