AudioPlayerViewController: Add new Audio Player
Add of an audio player view controller based on the player
view controller implementation.

Introducing a new audio player view with an improved user
experience.

The playqueue integration has been revisited has well as
the position of the player controls and the way to access
the audio options.

The audio options shown are now only the ones that are
linked to the audio sessions.

Closes #1276
Closes #1277