VideoPlayerViewController: Add play as audio feature
Add of a play as audio feature to allow users to play video files
as audio files.

When enabled, the video player will update the interface to showcase
a thumbnail of the current state of the media.

The minimized player will also follow the same behavior as a if an audio
file was being played.

Closes #554