MediaCategoryViewController: Fix album header appearance
Fix the status bar & navigation bar of the album view when the view
is loaded.

The correct theme is applied to the status bar to make it readable
no matter the theme and the position in the view.

The navigation bar's color is also updated when the view is loaded
taking into account the position in the view.

Closes #1481
Closes #1489