Split remote control from playback service
There is no need that the playback service is a delegate of the remote
control service when everything can be done within that one service.
Additionally, it should be owned by the app coordinator and not the
playback service.

This also fixes setting the playback position through a remote command.