AlbumModel: Remove the search call on the album tracks
The tracks are no longer checked in the Album model.

This check is already performed in the Media model with the album
track API and was leading to an infinite loop and therefore the crash
of the application.

Closes #1601

(cherry picked from commit 502e772d7d1743fa3bd6995c147e21a5eec18871)