VPS: remove hack that is no longer needed
VLCKit correctly handles events now so this is no longer needed and in
fact leads to failures when playing playlists with just a singular item.

This is a manual forward-port of 43338679 that was merged to the 3.3.x
branch only to address #1466. Fixes #1608

(cherry picked from commit 62547949748ad1b55ee78e180f392c02de33224c)