VLCPlayerDisplayController: Present over context + dismiss mini player when needed
Both the video and audio players are presented over the current context
meaning that when the minimizing gesture will be performed, the user
will be able to see the parent view controller instead of a black screen.

The mini player is also dismissed when the player enters the fullscreen
mode in order to make the animation smoother.