AudioPlayer: optimize metadata handling code
This removes custom logic to retrieve artist and title by trusting the
existing logic implemented for metadata handling.

Most importantly, this limits updates of the background color and the
artwork view to changes of the image as the metadata update function
can be called up to eight times per seek event leading to a serious
performance issue on older devices in case the artwork and the
background color are updated every time.

Fixes #1681