AudioPlayerView: Replace seek buttons by shuffle and repeat buttons
The seek buttons are here replaced by the shuffle and repeat mode ones
in order to make them more reachable.

The icons of these buttons are also updated to improve the accessibility.
When the shuffle mode is enabled and/or a repeat mode is selected, the respective
icons are updated and colored to make them stand out for the user.

An indicator in a form of a dot also appears below the icons when the options are
enabled for accessibility purposes.