QueueViewController: Improve the cells' animation when editing
When reorganizing the collection view, the cells are now properly placed
even if the gesture ends outside the bounds of the play queue.

The animation of the currently playing cell is properly reloaded after being
moved.

The scroll view is disabled for the currently playing cell without disabling
the user interaction of the entire cell.

Closes #1758