VLCNetworkServerBrowserViewController: Fix selection of searched files
When searching for medias when browsing a network share, the correct data
structure is selected in order to create the media list when necessary.

This means that only medias matching with the searched text will be added
to the media list when the search controller is active, otherwise every media
present in the collection will be added to the media list.

The automaticallyPlayNextItem setting is now also taken into consideration when
creating media lists with the search controller being active.

Closes #1451