VideoPlayerViewController: Update the aspect ratio action sheet
The aspect ratio action sheet is now a MediaPlayerActionSheet.

This allows the action sheet to be updated accordingly when the theme
of the application is updated and avoid any color conflicts.

In addition, a preconditionFailure is removed in order to allow touch
actions that does not involves a view opening to match the aspect ratio
action sheet's purpose.