VLCNetworkServerBrowserViewController: Favorite feature improvements
On iOS 13.0 and above, the NetworkServerBrowserViewController's table
view now has a long press contextual menu allowing to mark and unmark
folders as favorites.

This avoids displaying the heart shaped buttons on every row as we now
only display it for folders marked as favorites.

The previous behavior is applied for iOS versions below than 13.0.