VLCServerListViewController: Fix the local files playback
This fixes the several behaviours of the local files playback.

The user is now able to select one or multiple files and hit the "Open"
button in order to start the playback.

The user can also hit the "Open" button without having any media selected
in order to start the playback of the entire folder.