VideoPlayerViewController: Add notifications when adding or removing medias to the queue
Appending a media to the queue or removing one from it can now impact
the display of the player controls.

If the play queue contains a single media after such an action, the
player controls will display the seek buttons, if not the skip buttons
will be displayed allowing the user to select different medias from
the queue.