SettingsSection: Add a setting to force the rotation lock
This setting enables the user to automatically enable the rotation
lock option when the video player is displayed.

The orientation that will be locked is the one selected by the device
when the video player is displayed for the first time.
It remains the same until the option is disabled or the video player is
killed.

Closes #1869