MediaMoreOptionsActionSheet: Remove icon hack
This removes the hack in the options' action sheet that was displaying
the same icon for the bookmarks and chapter.

Since we do have a proper icon for the chapters, it is no longer needed.

Closes #1757

(cherry picked from commit b7aad04be6ddee434b1f3d1f24a9f83834c2d20c)