Use username & password from UI to log in to Plex
Currently Plex login ignores the username and password from the UI
and instead always looks up the Keychain for credentials. This will
fail if no credentials have been stored, or if the stored credentials
are incorrect. This will also result in unexpected behavior if the
user wishes to log in with different credentials on the same server.
This change means the code will attempt to log in with the given
username and password before falling back to credentials stored in the
Keychain.

Closes #1824

(cherry picked from commit 92bd954235dc6f1480f6ec1329bfa8193573a818)