TabBarCoordinator: Add a side tool bar
The side bar navigation now has a proper edit tool bar displaying the
edit actions.

All iPads running with iPadOS 18 or above now have the possibility to use
the regular edit tool bar at the bottom of the screen when the side bar is hidden
and the new side tool bar when the side bar navigation is displayed.