VLCPlayerDisplayController: Add the support of the arrow keys on the mini player
The mini player now supports the use of the left and right arrow keys to perform
a backwards and forwards seek.

The priority over system behaviour variable is set to true based on the API
documentation in order to match the behaviour of the same feature on both
the video and audio players.