MediaNavigationBar: Setup the rotation button only on iOS
Due to system limitations we cannot force a window's rotation of an
iPad application.

Any other hack also not a solution since it might break the AppStore
Connect validation or create an undefined behaviour.

Hence, the rotation button is removed from any iPad device but the
feature remains available on iOS.

Closes #1914